Further experimentation with materials
I suspect there is an issue when importing models from Blender, in that
a material’s texture is not applied to the model.  This issue doesn’t
happen when applying the same texture to UE’s primitive shapes.  The
normals on the Blender static mesh appear to be correct. Need to
investigate further.